refactor: extract UI components from Applications page
